# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDOLE AFRICA ADMIN - Unified operational management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List purchase orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /proformas:
    get:
      operationId: listProformas
      tags: [orders]
      summary: List proformas
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated proformas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProformas"
    post:
      operationId: createProforma
      tags: [orders]
      summary: Create a proforma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProformaBody"
      responses:
        "201":
          description: Created proforma
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proforma"
  /proformas/{id}:
    get:
      operationId: getProforma
      tags: [orders]
      summary: Get proforma details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proforma details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proforma"
    put:
      operationId: updateProforma
      tags: [orders]
      summary: Update a proforma
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProformaBody"
      responses:
        "200":
          description: Updated proforma
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proforma"
  /invoices:
    get:
      operationId: listInvoices
      tags: [orders]
      summary: List invoices
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInvoices"
    post:
      operationId: createInvoice
      tags: [orders]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [orders]
      summary: Get invoice details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [orders]
      summary: Update an invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /payments:
    get:
      operationId: listPayments
      tags: [orders]
      summary: List payments
      parameters:
        - in: query
          name: invoiceId
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPayments"
    post:
      operationId: createPayment
      tags: [orders]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [draft, confirmed, delivered, cancelled]
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referenceNumber
        - status
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        clientId:
          type: string
        status:
          type: string
          enum: [draft, confirmed, delivered, cancelled]
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: string
      required:
        - status
    Proforma:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        orderId:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [draft, sent, validated, rejected]
        totalAmount:
          type: number
        currency:
          type: string
        validUntil:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referenceNumber
        - status
        - createdAt
    CreateProformaBody:
      type: object
      properties:
        orderId:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: [draft, sent, validated, rejected]
        totalAmount:
          type: number
        currency:
          type: string
        validUntil:
          type: string
          format: date
        notes:
          type: string
      required:
        - status
    Invoice:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        proformaId:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        totalAmount:
          type: number
        paidAmount:
          type: number
        currency:
          type: string
        dueDate:
          type: string
          format: date
        issuedAt:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referenceNumber
        - status
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        proformaId:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        totalAmount:
          type: number
        currency:
          type: string
        dueDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - status
    Payment:
      type: object
      properties:
        id:
          type: string
        invoiceId:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [bank_transfer, cash, mobile_money, check, other]
        reference:
          type: string
        paidAt:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceId
        - amount
        - method
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        invoiceId:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [bank_transfer, cash, mobile_money, check, other]
        reference:
          type: string
        paidAt:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - invoiceId
        - amount
        - method
...
    PaginatedOrders:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    PaginatedProformas:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Proforma"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    PaginatedInvoices:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    PaginatedPayments:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── MESSAGING ─────────────────────────────────────────────────────────────
...